Load metrics in a single http get
Instead of executing a get per metric, just fetch them all in one go.
This avoids the performance issue associated when there are lots of
metrics. There is now a problem in theory that the single request could
grow so large that itself causes problems, we should look at
implementing paging in future to fetch the metrics in batches.

PNDA-3086